major refactoring
This updates the plugin to make use of the current standard.

All available policies have been updated accordingly and the old
X-Content-Security-Policy header has been dropped.

The actual code has been simplified to avoid duplicate code.

The enableHeader option has been dropped as it is the same as simply
disabling the plugin.